+++ example/lib/presentation/pages/dio_page.dart
@@ -28,11 +28,11 @@ class DioPage extends ChangeNotifierPage<DioProvider> {
               children: [
                 Text(
                   'Planet',
-                  style: Theme.of(context).textTheme.headline6,
+                  style: Theme.of(context).textTheme.titleLarge,
                 ),
                 Text(
                   'Create http activity using planet data',
-                  style: Theme.of(context).textTheme.bodyText2,
+                  style: Theme.of(context).textTheme.bodyMedium,
                 ),
               ],
             ),
@@ -42,7 +42,7 @@ class DioPage extends ChangeNotifierPage<DioProvider> {
             title: const Text('Get'),
             subtitle: Text(
               'Create http activity with method GET',
-              style: Theme.of(context).textTheme.caption,
+              style: Theme.of(context).textTheme.bodySmall,
             ),
             onTap: () {
               provider(context).fetchPlanet!.execute();
@@ -52,7 +52,7 @@ class DioPage extends ChangeNotifierPage<DioProvider> {
             title: const Text('Post'),
             subtitle: Text(
               'Create http activity with method POST',
-              style: Theme.of(context).textTheme.caption,
+              style: Theme.of(context).textTheme.bodySmall,
             ),
             onTap: () {
               provider(context).createPlanet!.execute(
@@ -65,7 +65,7 @@ class DioPage extends ChangeNotifierPage<DioProvider> {
             title: const Text('Put'),
             subtitle: Text(
               'Create http activity with method PUT',
-              style: Theme.of(context).textTheme.caption,
+              style: Theme.of(context).textTheme.bodySmall,
             ),
             onTap: () {},
           ),
@@ -73,7 +73,7 @@ class DioPage extends ChangeNotifierPage<DioProvider> {
             title: const Text('Delete'),
             subtitle: Text(
               'Create http activity with method DELETE',
-              style: Theme.of(context).textTheme.caption,
+              style: Theme.of(context).textTheme.bodySmall,
             ),
             onTap: () {},
           ),