+++ CHANGELOG.md
@@ -1,5 +1,8 @@
-## 1.1.0 - 15 May 2023
+## 1.1.0 - 7 June 2023
 
+* Upgrade Dependency with latest version
+* Solved breaking changes of dependency
+* Fix warning linter
 * Implement Http Activity filter by status code
 
 ## 1.0.4 - 1 February 2023